Apply changes requested by craftamap & Add KDoc compliant comments
