+++ app/src/androidTest/assets/tflite_negative_test.png
