+++ app/src/androidTest/assets/tflite_test.jpg
