+++ app/src/androidTest/java/de/hsaugsburg/teamulster/sohappy/detector/FerTFLiteSmileDetectorImplTest.kt
@@ -1,4 +1,4 @@
-package de.hsaugsburg.teamulster.sohappy
+package de.hsaugsburg.teamulster.sohappy.detector
 
 import android.content.Context
 import android.graphics.BitmapFactory
@@ -6,7 +6,8 @@ import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.launchActivity
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import de.hsaugsburg.teamulster.sohappy.analyzer.detector.TFLiteImpl
+import de.hsaugsburg.teamulster.sohappy.CameraActivity
+import de.hsaugsburg.teamulster.sohappy.analyzer.detector.FerTFLiteSmileDetectorImpl
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -17,8 +18,8 @@ import org.junit.runner.RunWith
  * See [testing documentation](http://d.android.com/tools/testing).
  */
 @RunWith(AndroidJUnit4::class)
-class SmileDetectorTest {
-    lateinit var tfliteImpl: TFLiteImpl
+class FerTFLiteSmileDetectorImplTest {
+    lateinit var tfliteImpl: FerTFLiteSmileDetectorImpl
     private lateinit var scenario: ActivityScenario<CameraActivity>
     private val instrumentationContext: Context =
         InstrumentationRegistry.getInstrumentation().context
@@ -26,13 +27,13 @@ class SmileDetectorTest {
     @Before
     fun setUp() {
         scenario = launchActivity<CameraActivity>()
-        scenario.onActivity { tfliteImpl = TFLiteImpl(it) }
+        scenario.onActivity { tfliteImpl = FerTFLiteSmileDetectorImpl(it) }
     }
 
     @Test
     fun useSmileDetectorWithPositiveResult() {
         scenario.onActivity {
-            val istr = instrumentationContext.assets.open("test.jpg")
+            val istr = instrumentationContext.assets.open("tflite_test.jpg")
             val detectResult = tfliteImpl.detect(BitmapFactory.decodeStream(istr))
             assert(detectResult.isSmiling)
         }
@@ -41,7 +42,7 @@ class SmileDetectorTest {
     @Test
     fun useSmileDetectorWithNegativeResult() {
         scenario.onActivity {
-            val istr = instrumentationContext.assets.open("negative_test.png")
+            val istr = instrumentationContext.assets.open("tflite_negative_test.png")
             val detectResult = tfliteImpl.detect(BitmapFactory.decodeStream(istr))
             assert(!detectResult.isSmiling)
         }
