+++ app/src/main/java/de/hsaugsburg/teamulster/sohappy/analyzer/detector/SmileDetector.kt
@@ -4,7 +4,10 @@ import android.graphics.Bitmap
 
 interface SmileDetector {
     companion object {
-        open class SmileDetectionResult (val isSmiling: Boolean)
+        interface SmileDetectionResult {
+            val isSmiling: Boolean
+            val predictionResults: ArrayList<AbstractTFLiteSmileDetector.Companion.Recognition>
+        }
     }
 
     val detectorName: String
